requirements.txt
cumberbatch             # testing
enum                    # util
GoogleAppEngineCloudStorageClient   # cloud_storage
nose                    # testing
python-dateutil         # cloud_storage
toolz                   # util
